Version 1.0.8.1

+++ src/crypto/hash.c
@@ -1,4 +1,4 @@
-// Copyright (c) 2012-2014, The CryptoNote developers, The Bytecoin developers
+// Copyright (c) 2012-2015, The CryptoNote developers, The Bytecoin developers
 //
 // This file is part of Bytecoin.
 //
@@ -27,7 +27,7 @@ void hash_permutation(union hash_state *state) {
 }
 
 void hash_process(union hash_state *state, const uint8_t *buf, size_t count) {
-  keccak1600(buf, count, (uint8_t*)state);
+  keccak1600(buf, (int)count, (uint8_t*)state);
 }
 
 void cn_fast_hash(const void *data, size_t length, char *hash) {
